refactor(checker): split get_class_instance_type_inner megafn into named phases (#13418)
