+++ crates/tsz-checker/src/types/class_type/mod.rs
@@ -5,6 +5,8 @@ mod core;
 mod entry;
 mod helpers;
 mod heritage_identity;
+mod instance;
+mod instance_merge;
 mod js_class_properties;
 mod prescan;
